refactor mocks and models
