+++ modules/raveboard/src/app/app.module.ts
@@ -4,7 +4,7 @@ import { NgModule } from '@angular/core';
 import { AppComponent } from './app.component';
 import { NodeComponent } from './svg/elements/node.component';
 import { ConnectorComponent } from './svg/elements/connector.component';
-import { MockDataControllerComponent } from './mock-data/mock-data-controller/mock-data-controller.component';
+import { MockDataControllerComponent } from './model/mocks/mock-data-controller.component';
 import { ActivationSpikeGraphComponent } from './svg/activation-spike-graph/activation-spike-graph.component';
 
 @NgModule({
