+++ modules/raveboard/src/app/model/message-from-ui.ts
@@ -1,0 +1,4 @@
+export interface MessageFromUI {
+    type: 'input',  // input message (from the UI to the ravestate server)
+    text: string	// message text
+}
