+++ modules/raveboard/src/app/model/message-to-ui.ts
@@ -1,0 +1,4 @@
+export interface MessageToUI {
+    type: 'output', // output message (from the ravestate server to the UI)
+    text: string	// message text
+}
